pyproject.toml
[project]
name = "simple-tg-bot-app"
version = "0.1.0"
description = ""
authors = [
    {name = "Andrey Usov",email = "nkt.andr@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi[all] (>=0.116.1,<0.117.0)",
    "aiogram (>=3.21.0,<4.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "sqladmin (>=0.21.0,<0.22.0)",
    "alembic (>=1.16.4,<2.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
